Replace Math.random() with Mersenne Twister as default PRNG
The PRNG algorithm behind V8’s Math.random() is called MWC1616. It is not very random at all, and is considered to be a poor quailty PRNG.

The Mersenne Twister, specifically the MT19937 implementation, provides fast generation of high-quality psuedorandom numbers. It will replace Math.random() as the default PRNG in DiceTower until a better option is found.